requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
